refactor: rename repository and repackage
