+++ src/main/java/com/lxp/auth/infrastructure/persistence/local/read/adapter/LocalAuthReadAdapter.java
@@ -2,7 +2,7 @@ package com.lxp.auth.infrastructure.persistence.local.read.adapter;
 
 import com.lxp.auth.application.port.required.AuthQueryPort;
 import com.lxp.auth.application.port.required.query.AuthView;
-import com.lxp.auth.infrastructure.persistence.local.read.repository.LocalAuthReadRepository;
+import com.lxp.auth.infrastructure.persistence.local.read.repository.LocalAuthJpaReadRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -12,16 +12,16 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class LocalAuthReadAdapter implements AuthQueryPort {
 
-    private final LocalAuthReadRepository localAuthReadRepository;
+    private final LocalAuthJpaReadRepository localAuthJpaReadRepository;
     private final LocalAuthReadMapper localAuthReadMapper;
 
     @Override
     public Optional<AuthView> findByLoginIdentifier(String loginIdentifier) {
-        return localAuthReadRepository.findByLoginIdentifier(loginIdentifier).map(localAuthReadMapper::toView);
+        return localAuthJpaReadRepository.findByLoginIdentifier(loginIdentifier).map(localAuthReadMapper::toView);
     }
 
     @Override
     public Optional<AuthView> findByUserId(String userId) {
-        return localAuthReadRepository.findByIdWithProjection(userId).map(localAuthReadMapper::toView);
+        return localAuthJpaReadRepository.findByIdWithProjection(userId).map(localAuthReadMapper::toView);
     }
 }
