+++ src/main/java/com/lxp/auth/infrastructure/persistence/local/write/adapter/LocalAuthWriteAdapter.java
@@ -3,7 +3,7 @@ package com.lxp.auth.infrastructure.persistence.local.write.adapter;
 import com.lxp.auth.application.port.required.AuthCommandPort;
 import com.lxp.auth.domain.common.exception.DuplicatedEmailException;
 import com.lxp.auth.domain.local.model.entity.LocalAuth;
-import com.lxp.auth.infrastructure.persistence.local.write.repository.LocalAuthWriteRepository;
+import com.lxp.auth.infrastructure.persistence.local.write.repository.LocalAuthJpaWriteRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -11,20 +11,20 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class LocalAuthWriteAdapter implements AuthCommandPort {
 
-    private final LocalAuthWriteRepository localAuthWriteRepository;
+    private final LocalAuthJpaWriteRepository localAuthJpaWriteRepository;
     private final LocalAuthWriteMapper localAuthWriteMapper;
 
     @Override
     public void save(LocalAuth localAuth) {
-        if (localAuthWriteRepository.existsByLoginIdentifier(localAuth.loginIdentifier())) {
+        if (localAuthJpaWriteRepository.existsByLoginIdentifier(localAuth.loginIdentifier())) {
             throw new DuplicatedEmailException();
         }
-        localAuthWriteRepository.save(localAuthWriteMapper.toEntity(localAuth));
+        localAuthJpaWriteRepository.save(localAuthWriteMapper.toEntity(localAuth));
     }
 
     @Override
     public void remove(LocalAuth localAuth) {
-        localAuthWriteRepository.deleteById(localAuth.getId().asString());
+        localAuthJpaWriteRepository.deleteById(localAuth.getId().asString());
     }
 
 }
