+++ src/main/java/com/lxp/auth/infrastructure/persistence/local/write/repository/LocalAuthJpaWriteRepository.java
@@ -3,7 +3,7 @@ package com.lxp.auth.infrastructure.persistence.local.write.repository;
 import com.lxp.auth.infrastructure.persistence.local.write.entity.LocalAuthJpaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalAuthWriteRepository extends JpaRepository<LocalAuthJpaEntity, String> {
+public interface LocalAuthJpaWriteRepository extends JpaRepository<LocalAuthJpaEntity, String> {
 
     boolean existsByLoginIdentifier(String loginIdentifier);
