+++ src/main/java/com/lxp/auth/infrastructure/web/external/client/adapter/UserServiceMapper.java
@@ -1,9 +1,9 @@
-package com.lxp.auth.infrastructure.web.external.adapter;
+package com.lxp.auth.infrastructure.web.external.client.adapter;
 
 import com.lxp.auth.application.port.required.command.CreateUserCommand;
 import com.lxp.auth.application.port.required.query.UserInfo;
-import com.lxp.auth.infrastructure.web.external.dto.CreateUserRequest;
-import com.lxp.auth.infrastructure.web.external.dto.UserResponse;
+import com.lxp.auth.infrastructure.web.external.client.dto.CreateUserRequest;
+import com.lxp.auth.infrastructure.web.external.client.dto.UserResponse;
 import org.springframework.stereotype.Component;
 
 @Component
