+++ src/main/java/com/lxp/auth/infrastructure/web/external/controller/LocalAuthExternalController.java
@@ -1,4 +1,4 @@
-package com.lxp.auth.infrastructure.web.internal;
+package com.lxp.auth.infrastructure.web.external.controller;
 
 import com.lxp.auth.application.port.provided.command.LocalAuthLogoutCommand;
 import com.lxp.auth.application.port.provided.usecase.LocalAuthLoginUseCase;
@@ -6,9 +6,8 @@ import com.lxp.auth.application.port.provided.usecase.LocalAuthLogoutUseCase;
 import com.lxp.auth.application.port.provided.usecase.LocalAuthRegisterUseCase;
 import com.lxp.auth.domain.common.constants.CookieConstants;
 import com.lxp.auth.domain.common.model.vo.AuthTokenInfo;
-import com.lxp.auth.infrastructure.web.internal.dto.LoginRequest;
-import com.lxp.auth.infrastructure.web.internal.dto.RegisterRequest;
-import com.lxp.auth.infrastructure.web.internal.mapper.LocalAuthInternalMapper;
+import com.lxp.auth.infrastructure.web.external.controller.dto.LoginRequest;
+import com.lxp.auth.infrastructure.web.external.controller.dto.RegisterRequest;
 import com.lxp.common.infrastructure.exception.ApiResponse;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -26,16 +25,16 @@ import java.util.Objects;
 @RestController
 @RequestMapping("/api-v1/auth")
 @RequiredArgsConstructor
-public class LocalAuthInternalController {
+public class LocalAuthExternalController {
 
     private final LocalAuthLoginUseCase localAuthLoginUseCase;
     private final LocalAuthRegisterUseCase localAuthRegisterUseCase;
     private final LocalAuthLogoutUseCase localAuthLogoutUseCase;
-    private final LocalAuthInternalMapper localAuthInternalMapper;
+    private final LocalAuthExternalMapper localAuthExternalMapper;
 
     @PostMapping("/login")
     public ResponseEntity<ApiResponse<Void>> login(@RequestBody LoginRequest request, HttpServletResponse response) {
-        AuthTokenInfo tokenInfo = localAuthLoginUseCase.execute(localAuthInternalMapper.localAuthLoginCommand(request));
+        AuthTokenInfo tokenInfo = localAuthLoginUseCase.execute(localAuthExternalMapper.localAuthLoginCommand(request));
         ResponseCookie cookie = createCookie(tokenInfo);
 
         response.addHeader(HttpHeaders.SET_COOKIE, cookie.toString());
@@ -45,7 +44,7 @@ public class LocalAuthInternalController {
 
     @PostMapping("/register")
     public ResponseEntity<ApiResponse<Void>> register(@RequestBody RegisterRequest request) {
-        localAuthRegisterUseCase.execute(localAuthInternalMapper.localAuthRegisterCommand(request));
+        localAuthRegisterUseCase.execute(localAuthExternalMapper.localAuthRegisterCommand(request));
         return ResponseEntity.ok(ApiResponse.success());
     }
