+++ src/main/java/com/lxp/auth/infrastructure/web/external/controller/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.lxp.auth.infrastructure.web.internal.dto;
+package com.lxp.auth.infrastructure.web.external.controller.dto;
 
 public record LoginRequest(
     String email, String password
